% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdata.R
\name{pcdata}
\alias{pcdata}
\title{progresively Censored Data}
\usage{
pcdata(m, r, qdistr = "qunif", par = c(0, 1))
}
\arguments{
\item{m}{vector of quantiles.}

\item{r}{vector of censoring scheme.}

\item{qdistr}{quantile function for specifided distribution.}

\item{par}{vector of  distribution parameters.}
}
\value{
\code{pcdata} gives the progressive censored data.
}
\description{
random generations under progressive censoring.
}
\details{
The details will be given.
}
\examples{
pcdata(10,rep(1,10),"qweibull",c(2,3))
}
\references{
Balakrishnan, N., & Aggarwala, R. (2000).
\emph{Progressive censoring: theory, methods, and applications}
Springer Science & Business Media.
}
